Corrects milestone and partner keys
These were commented out which is likely what is causing many of the errors seen on the dashboard. Another thing that is causing errors is the inclusion of non-18F team members in this list. I will file an issue in 18F/about_yml to attempt to support that but in the mean time, let's leave these team members commented so that we can quickly re-add them when it's possible.